refactor cfg visitor to use improved file visitor.
